+++ CHANGELOG.md
@@ -7,13 +7,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Keep format of highlighted track when it is playing - [#44](https://github.com/Rigellute/spotify-tui/pull/44) thanks to [@jfaltis](https://github.com/jfaltis)
+
 ## [0.0.5] - 2019-10-11
 
+### Added
+
 - Add `Ctrl-r` to cycle repeat mode ([@baxtea](https://github.com/baxtea))
-- Fix duplicate albums showing in artist discographies ([@baxtea](https://github.com/baxtea))
-- Slightly better error message with some debug tips when tracks fail to play
 - Refresh token when token expires ([@fangyi-zhou](https://github.com/fangyi-zhou))
 - Upgrade `rspotify` to fix [#39](https://github.com/Rigellute/spotify-tui/issues/39) ([@epwalsh](https://github.com/epwalsh))
+
+### Changed
+
+- Fix duplicate albums showing in artist discographies ([@baxtea](https://github.com/baxtea))
+- Slightly better error message with some debug tips when tracks fail to play
 
 ## [0.0.4] - 2019-10-05