+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Start a web server on localhost to display a simple webpage for the Redirect URI. Should hopefully improve the onboarding experience.
 - Add ability to skip to tracks using `n` for next and `p` for previous - thanks to [@samcal](https://github.com/samcal)
 - Implement seek functionality - you can now use `<` to seek backwards 5 seconds and `>` to go forwards 5 seconds
+- The event `A` will jump to the album list of the first artist in the track's artists list - closing [#45](https://github.com/Rigellute/spotify-tui/issues/45)
 
 ## [0.0.5] - 2019-10-11